Preliminary Exam Version (Fixed)
This commit contains the dissertation in the form it was after the preliminary exam. The first chapter is published (though the final version hasn't been transferred into the main document yet), and the second chapter is prepared for publication. The third chapter is incomplete.

Note: This version is fixed as the original commit has large files.